Extract pure component
