+++ src/cow-react/modules/wallet/web3-react/containers/Web3Status/index.tsx
@@ -1,55 +1,47 @@
-import styled from 'styled-components/macro'
-import { Web3StatusInner, Web3StatusConnected } from './Web3StatusMod'
 import { useWalletInfo, WalletModal } from '@cow/modules/wallet'
 import { STORAGE_KEY_LAST_PROVIDER } from 'constants/index'
 import { useCategorizeRecentActivity } from '@cow/common/hooks/useCategorizeRecentActivity'
 
-export const Wrapper = styled.div`
-  color: ${({ theme }) => theme.text1};
-  height: 40px;
-  width: 100%;
-  display: flex;
-  padding: 0;
-  margin: 0;
-  justify-content: center;
+import { useWeb3React } from '@web3-react/core'
+import { getConnection } from '@cow/modules/wallet/web3-react/utils/connection'
+import { useAppSelector } from 'state/hooks'
 
-  ${({ theme }) => theme.mediaWidth.upToMedium`
-    width: auto;
-    height: 100%;
-    margin: 0 0 0 auto;
-  `};
 
-  button {
-    height: auto;
-    border-radius: 21px;
-    padding: 6px 12px;
-    width: max-content;
-  }
-
-  ${Web3StatusConnected} {
-    height: 100%;
-    width: 100%;
-
-    > div > svg > path {
-      stroke: ${({ theme }) => theme.text3};
-    }
-  }
-`
+import { useToggleWalletModal } from 'state/application/hooks'
+import { useCloseFollowTxPopupIfNotPendingOrder } from '@src/custom/components/Popups/FollowPendingTxPopup'
+import { Web3StatusInner } from '@cow/modules/wallet/api/pure/Web3StatusInner'
+import { Wrapper } from '@cow/modules/wallet/api/pure/Web3StatusInner/styled'
 
 export function Web3Status() {
-  const walletInfo = useWalletInfo()
+  const { connector } = useWeb3React()
+  const { account, active, chainId, ensName } = useWalletInfo()
+  const connectionType = getConnection(connector).type
+
+  const error = useAppSelector((state) => state.connection.errorByConnectionType[getConnection(connector).type])
+
+  
+  const toggleWalletModal = useToggleWalletModal()
+  useCloseFollowTxPopupIfNotPendingOrder()
+
   const latestProvider = localStorage.getItem(STORAGE_KEY_LAST_PROVIDER)
 
   const { pendingActivity } = useCategorizeRecentActivity()
 
-  const { active } = walletInfo
   if (!active && !latestProvider) {
     return null
   }
 
   return (
     <Wrapper>
-      <Web3StatusInner pendingCount={pendingActivity.length} />
+      <Web3StatusInner
+        pendingCount={pendingActivity.length} 
+        account={account} 
+        chainId={chainId} 
+        error={error}
+        ensName={ensName}
+        connectWallet={toggleWalletModal}
+        connectionType={connectionType}
+      />
       <WalletModal />
     </Wrapper>
   )
